requirements.txt
pinecone
sentence-transformers
pypdf
python-dotenv
tqdm